+++ models/tenant/tenant.go
@@ -28,6 +28,9 @@ const (
 	// TenantMonitoredObjectType - datatype string used to identify a Tenant MonitoredObject in the datastore record
 	TenantMonitoredObjectType TenantDataType = "monitoredObject"
 
+	// ReportType - datatype string used to identify a Tenant Report in the datastore record
+	TenantReportType TenantDataType = "reports"
+
 	// TenantThresholdProfileType - datatype string used to identify a Tenant Ingestion Profile in the datastore record
 	TenantThresholdProfileType TenantDataType = "thresholdProfile"